Refactor: Zustand 파일 분리
