+++ src/store/useTestCodeStore.jsx
@@ -1,0 +1,8 @@
+import { create } from "zustand";
+
+const useTestCodeStore = create((set) => ({
+  testCodes: "",
+  setTestCodes: (testCodes) => set({ testCodes }),
+}));
+
+export default useTestCodeStore;
